+++ docs/changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 11/09/2024
 
+Version 0.0.6
+- 0.0.5 successfully tested on Flywheel
+- Included dynamic outlier thresholding in config
+
 Version 0.0.4
 - Debugged & successfully tested locally
 - Catches project run from